// types/park.ts
export type Park = {
    numberID: number;
    textOfficialName: string;
    textCommonName: string;
    textStatus: string;
    textType: string;
    textClass: string;
    textAddress: string;
    numberArea: number;
    numberLongitude: number;
    numberLatitude: number;
    distance?: number;
  };
  